split 'add job' button to separate component
